@@ -1,15 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <TargetLatestRuntimePatch>true</TargetLatestRuntimePatch>
-    <LangVersion>latest</LangVersion>
+    <Nullable>enable</Nullable>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     <Platforms>x64</Platforms>
+    <NoWarn>$(NoWarn);8618;NU1605</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Signum.Analyzer" Version="1.0.1" />
-    <PackageReference Include="Signum.MSBuildTask" Version="1.0.3" />
+    <PackageReference Include="Signum.Analyzer" Version="2.7.0" />
+    <PackageReference Include="Signum.MSBuildTask" Version="2.0.1" />
   </ItemGroup>
 
   <ItemGroup>
