@@ -1,20 +1,25 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <TypeScriptToolsVersion>3.2</TypeScriptToolsVersion>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TypeScriptToolsVersion>3.9</TypeScriptToolsVersion>
+    <TypeScriptBuildMode>true</TypeScriptBuildMode>
+    <TargetLatestRuntimePatch>true</TargetLatestRuntimePatch>
+    <Nullable>enable</Nullable>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     <Platforms>x64</Platforms>
     <IsPackable>false</IsPackable>
-    <LangVersion>latest</LangVersion>
-    <TargetLatestRuntimePatch>true</TargetLatestRuntimePatch>
+    <AssemblyVersion>1.0.0.*</AssemblyVersion>
+    <FileVersion>1.0.0.0</FileVersion>
+    <Deterministic>false</Deterministic>
     <UserSecretsId>Southwind</UserSecretsId>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <NoWarn>$(NoWarn);1591</NoWarn>
+    <NoWarn>$(NoWarn);1591;NU1605</NoWarn>
+    <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <OutputPath>bin\</OutputPath>
-    <NoWarn>1701;1702,1591</NoWarn>
   </PropertyGroup>
   
   <ItemGroup>
@@ -26,29 +31,38 @@
     <Compile Remove="CodeGen\**" />
     <Compile Remove="dist\**" />
     <Compile Remove="Translations\**" />
+    <Compile Remove="ts_out\**" />
     <Content Remove="CodeGen\**" />
     <Content Remove="dist\**" />
     <Content Remove="Translations\**" />
+    <Content Remove="ts_out\**" />
     <EmbeddedResource Remove="CodeGen\**" />
     <EmbeddedResource Remove="dist\**" />
     <EmbeddedResource Remove="Translations\**" />
+    <EmbeddedResource Remove="ts_out\**" />
     <None Remove="CodeGen\**" />
     <None Remove="dist\**" />
     <None Remove="Translations\**" />
+    <None Remove="ts_out\**" />
     <TypeScriptCompile Include="**\*.tsx" />
     <TypeScriptCompile Include="**\*.ts" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="CNTK.CPUOnly" Version="2.6.0" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Signum.Analyzer" Version="1.0.1" />
-    <PackageReference Include="Microsoft.TypeScript.MSBuild" Version="3.2.3">
+    <TypeScriptCompile Remove="node_modules\**" />
+    <TypeScriptCompile Remove="ts_out\**" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="CNTK.CPUOnly" Version="2.7.0" />
+    <PackageReference Include="Microsoft.TypeScript.MSBuild" Version="3.9.2">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="Signum.TSGenerator" Version="1.0.3" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
+    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.10.8" />
+    <PackageReference Include="Signum.Analyzer" Version="2.7.0" />
+    <PackageReference Include="Signum.TSGenerator" Version="3.0.0" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.4.1" />
   </ItemGroup>
 
   <ItemGroup>
