@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
@@ -12,14 +12,13 @@ namespace Southwind.Entities
     [Serializable, PrimaryKey(typeof(Guid))]
     public abstract class CustomerEntity : Entity
     {
-        [NotNullValidator]
         public AddressEmbedded Address { get; set; }
 
-        [StringLengthValidator(AllowNulls = false, Min = 3, Max = 24), TelephoneValidator]
+        [StringLengthValidator(Min = 3, Max = 24), TelephoneValidator]
         public string Phone { get; set; }
 
-        [StringLengthValidator(AllowNulls = true, Min = 3, Max = 24), TelephoneValidator]
-        public string Fax { get; set; }
+        [StringLengthValidator(Min = 3, Max = 24), TelephoneValidator]
+        public string? Fax { get; set; }
 
         public static readonly SessionVariable<CustomerEntity> CurrentCustomerVariable = Statics.SessionVariable<CustomerEntity>("Customer");
         public static CustomerEntity Current
@@ -39,21 +38,21 @@ namespace Southwind.Entities
     [Serializable, EntityKind(EntityKind.Shared, EntityData.Transactional)]
     public class PersonEntity : CustomerEntity
     {
-        [StringLengthValidator(AllowNulls = false, Min = 3, Max = 40)]
+        [StringLengthValidator(Min = 3, Max = 40)]
         public string FirstName { get; set; }
 
-        [StringLengthValidator(AllowNulls = false, Min = 3, Max = 40)]
+        [StringLengthValidator(Min = 3, Max = 40)]
         public string LastName { get; set; }
 
-        [StringLengthValidator(AllowNulls = true, Min = 3, Max = 10)]
-        public string Title { get; set; }
+        [StringLengthValidator(Min = 3, Max = 10)]
+        public string? Title { get; set; }
 
         [DateTimePrecisionValidator(DateTimePrecision.Days)]
         public DateTime? DateOfBirth { get; set; }
 
         public bool Corrupt { get; set; }
 
-        public override Dictionary<Guid, IntegrityCheck> EntityIntegrityCheck()
+        public override Dictionary<Guid, IntegrityCheck>? EntityIntegrityCheck()
         {
             using (this.Corrupt ? Corruption.AllowScope() : null)
             {
@@ -68,11 +67,6 @@ namespace Southwind.Entities
             {
                 this.Corrupt = false;
             }
-        }
-
-        protected override void PostRetrieving()
-        {
-            base.PostRetrieving();
         }
 
         public override string ToString()
@@ -90,21 +84,17 @@ namespace Southwind.Entities
     [Serializable, EntityKind(EntityKind.Shared, EntityData.Transactional)]
     public class CompanyEntity : CustomerEntity
     {
-        [StringLengthValidator(AllowNulls = true, Min = 3, Max = 40)]
+        [StringLengthValidator(Min = 3, Max = 40)]
         public string CompanyName { get; set; }
 
-        [StringLengthValidator(AllowNulls = false, Min = 3, Max = 30)]
+        [StringLengthValidator(Min = 3, Max = 30)]
         public string ContactName { get; set; }
 
-        [StringLengthValidator(AllowNulls = false, Min = 3, Max = 30)]
+        [StringLengthValidator(Min = 3, Max = 30)]
         public string ContactTitle { get; set; }
 
-        static Expression<Func<CompanyEntity, string>> ToStringExpression = e => e.CompanyName;
-        [ExpressionField]
-        public override string ToString()
-        {
-            return ToStringExpression.Evaluate(this);
-        }
+        [AutoExpressionField]
+        public override string ToString() => As.Expression(() => CompanyName);
     }
 
     [AutoInit]
