@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -9,14 +9,14 @@ namespace Southwind.Entities
     [Serializable]
     public class UserEmployeeMixin : MixinEntity
     {
-        protected UserEmployeeMixin(Entity mainEntity, MixinEntity next)
+        protected UserEmployeeMixin(ModifiableEntity mainEntity, MixinEntity next)
             : base(mainEntity, next)
         {
         }
 
         public AllowLogin AllowLogin { get; set; }
 
-        public Lite<EmployeeEntity> Employee { get; set; }
+        public Lite<EmployeeEntity>? Employee { get; set; }
     }
 
     public enum AllowLogin
