@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -16,7 +16,7 @@ namespace Southwind.Logic
 {
     public static class ProductLogic
     {
-        public static ResetLazy<Dictionary<CategoryEntity, List<ProductEntity>>> ActiveProducts;
+        public static ResetLazy<Dictionary<CategoryEntity, List<ProductEntity>>> ActiveProducts = null!;
 
         public static void Start(SchemaBuilder sb)
         {
@@ -44,7 +44,7 @@ namespace Southwind.Logic
                     Database.Query<ProductEntity>()
                     .Where(a => !a.Discontinued)
                     .Select(p => new { Category = p.Category.Entity, Product = p })
-                    .GroupToDictionary(a => a.Category, a => a.Product),
+                    .GroupToDictionary(a => a.Category!, a => a.Product!), /*CSBUG*/
                     new InvalidateWith(typeof(ProductEntity)));
 
                 QueryLogic.Queries.Register(ProductQuery.CurrentProducts, () =>
