@@ -5,24 +5,21 @@ import * as Finder from '@framework/Finder';
 import { PredictorEntity } from '@extensions/MachineLearning/Signum.Entities.MachineLearning';
 import SalesEstimation from './SalesEstimation';
 
-export default class Product extends React.Component<{ ctx: TypeContext<ProductEntity> }> {
-
-  render() {
-    const ctx = this.props.ctx;
-    return (
-      <div>
-        <ValueLine ctx={ctx.subCtx(p => p.productName)} />
-        <EntityCombo ctx={ctx.subCtx(p => p.supplier)} />
-        <EntityLine ctx={ctx.subCtx(p => p.category)} />
-        <ValueLine ctx={ctx.subCtx(p => p.quantityPerUnit)} />
-        <ValueLine ctx={ctx.subCtx(p => p.unitPrice)} />
-        <ValueLine ctx={ctx.subCtx(p => p.unitsInStock)} />
-        <SalesEstimation ctx={ctx} />
-        <ValueLine ctx={ctx.subCtx(p => p.reorderLevel)} />
-        <ValueLine ctx={ctx.subCtx(p => p.discontinued)} />
-        <EntityTable ctx={ctx.subCtx(p => p.additionalInformation)} />
-      </div>
-    );
-  }
+export default function Product(p : { ctx: TypeContext<ProductEntity> }){
+  const ctx = p.ctx;
+  return (
+    <div>
+      <ValueLine ctx={ctx.subCtx(p => p.productName)} />
+      <EntityCombo ctx={ctx.subCtx(p => p.supplier)} />
+      <EntityLine ctx={ctx.subCtx(p => p.category)} />
+      <ValueLine ctx={ctx.subCtx(p => p.quantityPerUnit)} />
+      <ValueLine ctx={ctx.subCtx(p => p.unitPrice)} />
+      <ValueLine ctx={ctx.subCtx(p => p.unitsInStock)} />
+      <SalesEstimation ctx={ctx} />
+      <ValueLine ctx={ctx.subCtx(p => p.reorderLevel)} />
+      <ValueLine ctx={ctx.subCtx(p => p.discontinued)} />
+      <EntityTable ctx={ctx.subCtx(p => p.additionalInformation)} />
+    </div>
+  );
 }
 
