@@ -1,13 +1,15 @@
-﻿require("react");
+//third-pary libs
+require("react");
 require("react-dom");
 require("prop-types");
 require("react-router");
 require("react-widgets");
-require("react-popper");
-require("react-overlays");
-require("query-string");
+require("react-bootstrap");
 require("moment");
 require("numbro");
-require("moment-duration-format");
-require("d3");
+require("@fortawesome/fontawesome-svg-core");
+require("@fortawesome/free-regular-svg-icons");
+require("@fortawesome/free-solid-svg-icons");
+require("@fortawesome/react-fontawesome");
+//require("d3");
 
