@@ -1,5 +1,6 @@
-﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Signum.React.Filters;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,7 +10,7 @@ namespace Southwind.React.Controllers
     public class HomeController : Controller
     {
         // GET: Default
-        [AllowAnonymous]
+        [SignumAllowAnonymous]
         public ActionResult Index()
         {
             return View();
